avutil/pixfmt: Add NV15, NV20BS and NV30
Add NV15 and NV20 pixel formats used by the Rockchip Video Decoder for
10-bit buffers.

NV15 and NV20 is 10-bit 4:2:0/4:2:2 semi-planar YUV formats similar to
NV12 and NV16, using 10-bit components with no padding between each
component. Instead, a group of 4 luminance/chrominance samples are
stored over 5 bytes in little endian order:

YYYY = UVUV = 4 * 10 bits = 40 bits = 5 bytes

The '15' and '20' suffix refers to the optimum effective bits per pixel
which is achieved when the total number of luminance samples is a
multiple of 8 for NV15 and 4 for NV20.

The existing NV20LE/BE pixdesc indicate that the existing format is not
packed and instead have more in common with P210, each 10-bit sample is
LSB aligned instead of MSB aligned.

Instead a new similar name, NV20BS (bitstream), was chosen.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>
---
Q: Should all be named NVxxBS or use some other naming convention ?
Link: https://gitlab.freedesktop.org/mesa/drm/-/merge_requests/329